Task1416-FixPackages-name
Task: https://github.com/jdi-testing/jdi-light/issues/1416
