+++ jdi-light-bdd-no-po-tests/src/test/java/cucmbertests/Runner.java
@@ -1,8 +1,8 @@
-package cucmberTests;
+package cucmbertests;
 
 import com.epam.jdi.light.elements.composite.Form;
 import com.epam.jdi.light.elements.composite.WebPage;
-import cucmberTests.steps.User;
+import cucmbertests.steps.User;
 import cucumber.api.CucumberOptions;
 import cucumber.api.junit.Cucumber;
 import cucumber.api.testng.AbstractTestNGCucumberTests;
@@ -16,7 +16,7 @@ import static com.epam.jdi.light.elements.init.UIFactory.$;
 @RunWith(Cucumber.class)
 @CucumberOptions(
     features = "classpath:features"
-    , glue = {"com.epam.jdi.bdd", "cucmberTests"}
+    , glue = {"com.epam.jdi.bdd", "cucmbertests"}
     //, tags = {"@fail-test"}
 )
 public class Runner extends AbstractTestNGCucumberTests {
