+++ jdi-light-bootstrap-tests/src/test/java/io/github/epam/bootstrap/tests/composite/section/buttongroup/ToolbarTests.java
@@ -1,4 +1,4 @@
-package io.github.epam.bootstrap.tests.composite.section.buttonGroup;
+package io.github.epam.bootstrap.tests.composite.section.buttongroup;
 
 import io.github.epam.TestsInit;
 import org.testng.annotations.BeforeMethod;
