+++ jdi-light-examples/src/main/java/selenium/LoginForm.java
@@ -1,4 +1,4 @@
-package selenium.seleniumPO;
+package selenium;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
