+++ jdi-light-examples/src/main/java/selenium/SiteSelenium.java
@@ -1,4 +1,4 @@
-package selenium.seleniumPO;
+package selenium;
 
 import selenium.entities.User;
