+++ jdi-light-examples/src/test/java/io/github/epam/tests/selenium/SeleniumPerformanceTests.java
@@ -6,9 +6,9 @@ import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 import selenium.entities.User;
 
-import static selenium.seleniumPO.SiteSelenium.homePage;
-import static selenium.seleniumPO.SiteSelenium.login;
-import static selenium.seleniumPO.SiteSelenium.performancePage;
+import static selenium.SiteSelenium.homePage;
+import static selenium.SiteSelenium.login;
+import static selenium.SiteSelenium.performancePage;
 
 
 public class SeleniumPerformanceTests {
